pyproject.toml
[project]
name = "dfc-shazam"
version = "0.1.0"
description = "MCP server for Dockerfile to Chainguard conversion assistance"
requires-python = ">=3.12"
dependencies = [
    "mcp[cli]>=1.0.0",
    "httpx>=0.27.0",
    "pydantic>=2.0.0",
    "pydantic-settings>=2.0.0",
    "pyyaml>=6.0.3",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "pytest-httpx>=0.30.0",
    "ruff>=0.4.0",
    "mypy>=1.10.0",
    "pyyaml>=6.0.0",
]

[project.scripts]
dfc-shazam = "dfc_shazam.server:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/dfc_shazam"]

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "I", "UP"]

[tool.mypy]
python_version = "3.12"
strict = true

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
markers = [
    "slow: marks tests as slow (docker builds, deselect with '-m \"not slow\"')",
    "docker: marks tests that require Docker to be running",
    "chainctl: marks tests that require chainctl authentication",
    "e2e: marks end-to-end tests (full conversion pipeline)",
]
addopts = ["-v", "--strict-markers"]
filterwarnings = [
    "ignore::DeprecationWarning",
]
